requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
cryptography==2.0.3
dnspython==1.15.0
idna==2.6
itsdangerous==0.24
pycparser==2.18
six==1.10.0
